Fix serializer classes not being inherited
This fixes an issue where the serializer classes would not be inherited
from the event's parent(s). This forced developers to specify the
`payload_serializer` and `customer_properties_serializer` every time
they inherited from `SolidusTracking::Event::Base` or one of its children.

The new implementation looks up the ancestors chain when computing
serializer classes, which is the expected behavior.