Update order serializer to handle promotions correctly
Previously, the order serializer would pull in all promotions regardless
of whether they were eligible or not and display the promotion_code.
This also causes breakage with order promotions that apply to all orders,
as those promotions do not have promotion_codes.

This only displays eligible promotions in the serializer, and if no
promotion code exists, will set the promotion code to "all orders".