Bump supported Solidus versions
We want to support Solidus 4 to 4.5.

I stopped at 4.5 because there are changes in 4.6+ with some
'disable_builtin_emails' configuration that required more work, so I
think a plan to move forward is to make the current gemspec only allow
'>= 4.0.0', '< 4.6', and add some deprecation warnings about that
feature becoming obsolete. Then have a new version that's '>= 4.6.0',
'<5.0' that removes it all.

I am doing this work specifically to bump another app that uses it to
Solidus 4, and that is my priority at the moment.