Set sqlite3 gem constraint
To fix an error we got,
```
Error loading the 'sqlite3' Active Record adapter. Missing a gem it
depends on? can't activate sqlite3 (~> 1.4), already activated
sqlite3-2.7.4-arm64-darwin.  Make sure all dependencies are added to
Gemfile.
```

The error message shows the version it was erroring to activate. Locking
it to that version and `bundle update sqlite3` fixed the issue.